# src/backend/internal/server/v2/docs/openapi.yaml
openapi: '3.0.2'
info:
  title: Спортивные залы
  version: v2
...
paths:
  /api/v2/gym:
    post:
      operationId: postGym
      tags:
        - Gym
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Gym'
      responses:
        '201':
          description: "Зал создан"
        '400':
          description: "Некорректный запрос"  
        '401':
          description: "Неавтризован"
        '500':
          description: "Внутренняя ошибка сервера"
    get:
      operationId: getListGyms
      tags:
        - Gym
      responses:
        '200':
          description: "Залы получены"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Gym'
        '404':
          description: "Залы не найдены"  
        '500':
          description: "Внутренняя ошибка сервера"      
  /api/v2/gym/{gymId}:
    get:
      operationId: getGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Зал получен"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Gym'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Зал не найден"  
        '500':
          description: "Внутренняя ошибка сервера"
    put:
      operationId: putGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutGym'
      responses:
        '204':
          description: "Зал обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Зал не найден"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteGym
      tags:
        - Gym
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Зал удален"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Зал не найден"
        '500':
          description: "Внутренняя ошибка сервера"
  /api/v2/gym/{gymId}/equipment:
    get:
      operationId: getListEquipments
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Оборудование получено"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Оборудование не найдено"  
        '500':
          description: "Внутренняя ошибка сервера"
    post:
      operationId: postEquipment
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipment'
      responses:
        '201':
          description: "Оборудование создано"
        '400':
          description: "Некорректный запрос"  
        '401':
          description: "Неавтризован"
        '500':
          description: "Внутренняя ошибка сервера"
  /api/v2/gym/{gymId}/equipment/{equipmentId}:      
    put:
      operationId: putEquipment
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: equipmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutEquipment'
      responses:
        '204':
          description: "Оборудование обновлено"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteEquipment
      tags:
        - Equipment
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: equipmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Оборудование удалено"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
  /api/v2/gym/{gymId}/membership_type:
    get:
      operationId: getListMembershipTypes
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid  
      responses:
        '200':
          description: "Типы абонементов получены"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MembershipType'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Абонементы не найдены"  
        '500':
          description: "Внутренняя ошибка сервера"
    post:
      operationId: postMembershipType
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipType'
      responses:
        '201':
          description: "Типа абонмента создан"
        '400':
          description: "Некорректный запрос"  
        '401':
          description: "Неавтризован"
        '500':
          description: "Внутренняя ошибка сервера"
  /api/v2/gym/{gymId}/membership_type/{membershipTypeId}:        
    put:
      operationId: putMembershipType
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutMembershipType'
      responses:
        '204':
          description: "Абонемент обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Оборудование не найдено"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteMembershipType
      tags:
        - MembershipType
      parameters:
        - name: gymId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: membershipTypeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Абонемент удален"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Абонемент не найден"
        '500':
          description: "Внутренняя ошибка сервера"
...
components:
  schemas:
    Gym:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор спортзала (UUID)"
        name:
          type: string
          description: "Название спортзала"
        phone:
          type: string
          description: "Телефон спортзала"
        city:
          type: string
          description: "Город, где находится спортзал"
        addres:
          type: string
          description: "Адрес спортзала"
        ischain:
          type: boolean
          description: "Флаг, указывающий, является ли спортзал частью сети"
    PutGym:
      type: object
      required:
        - name
        - phone
        - city
        - addres
        - ischain
      properties:
        name:
          type: string
          description: "Название спортзала"
        phone:
          type: string
          description: "Телефон спортзала"
        city:
          type: string
          description: "Город, где находится спортзал"
        addres:
          type: string
          description: "Адрес спортзала"
        ischain:
          type: boolean
          description: "Флаг, указывающий, является ли спортзал частью сети"
...
    Equipment:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор оборудования в формате UUID"
        name:
          type: string
          description: "Название оборудования"
        description:
          type: string
          description: "Описание оборудования"
    PutEquipment:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: "Название оборудования"
        description:
          type: string
          description: "Описание оборудования"
    MembershipType:
      type: object
      required:
        - id
        - type
        - description
        - price
        - daysduration
        - gymid
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор типа членства в формате UUID"
        type:
          type: string
          description: "Тип членства"
        description:
          type: string
          description: "Описание типа членства"
        price:
          type: string
          description: "Цена членства"
        daysduration:
          type: integer
          description: "Количество дней действия членства"
        gymid:
          type: string
          format: uuid
          description: "Уникальный идентификатор спортзала, к которому относится тип членства в формате UUID"
    PutMembershipType:
      type: object
      required:
        - type
        - description
        - price
        - daysduration
      properties:
        type:
          type: string
          description: "Тип членства"
        description:
          type: string
          description: "Описание типа членства"
        price:
          type: string
          description: "Цена членства"
        daysduration:
          type: integer
          description: "Количество дней действия членства"
...
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: bearer
      bearerFormat: JWT